Changed coef to double

+++ INCLUDE/DIAGS/diagram.h
@@ -12,10 +12,10 @@ template <class QL_Type>
 {
 	public:
 		std::vector<Trace<QL_Type>> traces;
-		mutable int coef;
+		mutable double coef;
 
 	Diagram(){coef=0;};
-	Diagram(int c, std::vector<Trace<QL_Type>> t):coef(c), traces(t){};
+	Diagram(double c, std::vector<Trace<QL_Type>> t):coef(c), traces(t){};
 
 	std::string name() const;
